Move text and image data into separate files, delete unused data
