+++ cardigan/stories/components/BookPromo/BookPromo.stories.tsx
@@ -1,9 +1,6 @@
-import {
-  image,
-  imagesBaseUrl,
-  singleLineOfText,
-} from '@weco/cardigan/stories/content';
 import BookPromo from '@weco/content/components/BookPromo/BookPromo';
+import { bookImageUrl, image } from '@weco/cardigan/stories/data/images';
+import { singleLineOfText } from '@weco/cardigan/stories/data/text';
 
 const Template = args => <BookPromo {...args} />;
 
@@ -11,7 +8,7 @@ export const basic = Template.bind({});
 basic.args = {
   book: {
     id: 'cardigan',
-    cover: image(`${imagesBaseUrl}/book.png`, 575, 884),
+    cover: image(bookImageUrl, 575, 884),
     title: singleLineOfText(),
     subtitle: singleLineOfText(),
   },
