+++ cardigan/stories/components/CaptionedImage/CaptionedImage.stories.tsx
@@ -1,5 +1,5 @@
 import CaptionedImage from '@weco/content/components/CaptionedImage/CaptionedImage';
-import { captionedImage } from '@weco/cardigan/stories/content';
+import { captionedImage } from '@weco/cardigan/stories/data/images';
 
 const Template = args => <CaptionedImage {...args} />;
 export const basic = Template.bind({});
