+++ cardigan/stories/components/Cards/Cards.stories.tsx
@@ -6,17 +6,16 @@ import ExhibitionPromo from '@weco/content/components/ExhibitionPromo/Exhibition
 import StoryPromo from '@weco/content/components/StoryPromo/StoryPromo';
 import { exhibitionBasic } from '@weco/cardigan/stories/data/prismic/exhibition';
 import PrismicImage from '@weco/common/views/components/PrismicImage/PrismicImage';
-import {
-  squareImage,
-  singleLineOfText,
-  bannerCardItem,
-  articleBasic,
-  image,
-  event,
-} from '@weco/cardigan/stories/content';
 import Readme from '@weco/content/components/FeaturedCard/README.md';
 import { ReadmeDecorator } from '@weco/cardigan/config/decorators';
 import { font } from '@weco/common/utils/classnames';
+import {
+  bannerCardItem,
+  articleBasic,
+  event,
+} from '@weco/cardigan/stories/data/content';
+import { image, squareImage } from '@weco/cardigan/stories/data/images';
+import { singleLineOfText } from '@weco/cardigan/stories/data/text';
 
 const primaryLabelList = [{ text: 'Study day' }, { text: 'Schools' }];
 const secondaryLabelList = [{ text: 'Speech-to-text' }];
