+++ cardigan/stories/data/prismic/exhibition.ts
@@ -1,4 +1,4 @@
-import { florenceWinterfloodImage } from '../../content';
+import { florenceWinterfloodImageUrl } from '../images';
 
 const exhibitionDocument = {
   id: 'YKTDpxAAACEAGIb7',
@@ -337,7 +337,7 @@ const exhibitionDocument = {
             alt: "Photograph of a young woman wearing a face covering exploring an exhibition gallery space. The woman is standing to the right of frame, arms clasped across her body, reading an information panel on the front of a large circular wooden board. This text is hidden from view as the camera is looking at the back of this wooden board. Behind the board is a large TV screen mounted to a vertical rectangular board wich is covered in a blue patterned upper section and a re patterned lower section. On the screen is a still from a video showing a demonstration or protest with people carrying banner. One of the banners has the words, 'No justice, no peace, no racist police!' written on it. Behind the young woman the gallery space recedes into the distance, with further exhibits and two large windows.",
             copyright:
               'What does it mean to be human now? Four views by CYP x CALLY | Photo: Steven Pocock | Wellcome Collection | | CC-BY-NC ||',
-            url: `${florenceWinterfloodImage}-3200x1800.jpg`,
+            url: florenceWinterfloodImageUrl('3200x1800'),
             '32:15': {
               dimensions: {
                 width: 3200,
@@ -346,7 +346,7 @@ const exhibitionDocument = {
               alt: "Photograph of a young woman wearing a face covering exploring an exhibition gallery space. The woman is standing to the right of frame, arms clasped across her body, reading an information panel on the front of a large circular wooden board. This text is hidden from view as the camera is looking at the back of this wooden board. Behind the board is a large TV screen mounted to a vertical rectangular board wich is covered in a blue patterned upper section and a re patterned lower section. On the screen is a still from a video showing a demonstration or protest with people carrying banner. One of the banners has the words, 'No justice, no peace, no racist police!' written on it. Behind the young woman the gallery space recedes into the distance, with further exhibits and two large windows.",
               copyright:
                 'What does it mean to be human now? Four views by CYP x CALLY | Photo: Steven Pocock | Wellcome Collection | | CC-BY-NC ||',
-              url: `${florenceWinterfloodImage}-3200x1500.jpg`,
+              url: florenceWinterfloodImageUrl('3200x1500'),
             },
             '16:9': {
               dimensions: {
@@ -356,7 +356,7 @@ const exhibitionDocument = {
               alt: "Photograph of a young woman wearing a face covering exploring an exhibition gallery space. The woman is standing to the right of frame, arms clasped across her body, reading an information panel on the front of a large circular wooden board. This text is hidden from view as the camera is looking at the back of this wooden board. Behind the board is a large TV screen mounted to a vertical rectangular board wich is covered in a blue patterned upper section and a re patterned lower section. On the screen is a still from a video showing a demonstration or protest with people carrying banner. One of the banners has the words, 'No justice, no peace, no racist police!' written on it. Behind the young woman the gallery space recedes into the distance, with further exhibits and two large windows.",
               copyright:
                 'What does it mean to be human now? Four views by CYP x CALLY | Photo: Steven Pocock | Wellcome Collection | | CC-BY-NC ||',
-              url: `${florenceWinterfloodImage}-3200x1800.jpg`,
+              url: florenceWinterfloodImageUrl('3200x1800'),
             },
             square: {
               dimensions: {
@@ -366,7 +366,7 @@ const exhibitionDocument = {
               alt: "Photograph of a young woman wearing a face covering exploring an exhibition gallery space. The woman is standing to the right of frame, arms clasped across her body, reading an information panel on the front of a large circular wooden board. This text is hidden from view as the camera is looking at the back of this wooden board. Behind the board is a large TV screen mounted to a vertical rectangular board wich is covered in a blue patterned upper section and a re patterned lower section. On the screen is a still from a video showing a demonstration or protest with people carrying banner. One of the banners has the words, 'No justice, no peace, no racist police!' written on it. Behind the young woman the gallery space recedes into the distance, with further exhibits and two large windows.",
               copyright:
                 'What does it mean to be human now? Four views by CYP x CALLY | Photo: Steven Pocock | Wellcome Collection | | CC-BY-NC ||',
-              url: `${florenceWinterfloodImage}-3200x3200.jpg`,
+              url: florenceWinterfloodImageUrl('3200x3200'),
             },
           },
           link: null,
@@ -407,7 +407,7 @@ export const exhibitionBasic = {
   title: 'What does it mean to be human now? Four views by CYP x CALLY',
   promo: {
     image: {
-      contentUrl: `${florenceWinterfloodImage}-3200x1800.jpg`,
+      contentUrl: florenceWinterfloodImageUrl('3200x1800'),
       width: 3200,
       height: 1800,
     },
